added most of vanilla recipes.

+++ diorite-api/src/main/java/org/diorite/material/items/DiamondLeggingsMat.java
@@ -179,7 +179,7 @@ public class DiamondLeggingsMat extends LeggingsMat
      */
     public static void register(final DiamondLeggingsMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((short) element.getType(), element);
         byName.put(element.getTypeName(), element);
     }
