+++ diorite-api/src/main/java/org/diorite/material/items/DiamondPickaxeMat.java
@@ -179,7 +179,7 @@ public class DiamondPickaxeMat extends PickaxeMat
      */
     public static void register(final DiamondPickaxeMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((short) element.getType(), element);
         byName.put(element.getTypeName(), element);
     }
