+++ diorite-api/src/main/java/org/diorite/material/items/WritableBookMat.java
@@ -117,7 +117,7 @@ public class WritableBookMat extends ItemMaterialData
      */
     public static void register(final WritableBookMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((short) element.getType(), element);
         byName.put(element.getTypeName(), element);
     }
